refactor: separates the default interface
- create unsafe Schnorrkel
- export UnsafeSchnorrkel
- remove fromJson and toJson from the Schnorrkel
- modify tests
