+++ src/index.ts
@@ -1,4 +1,5 @@
 import Schnorrkel from './schnorrkel'
+export { default as UnsafeSchnorrkel } from './unsafe-schnorrkel'
 
 export { Key, KeyPair, Signature, PublicNonces, Challenge, SignatureOutput, FinalPublicNonce } from './types'
