+++ src/schnorrkel.ts
@@ -1,13 +1,13 @@
 import secp256k1 from 'secp256k1'
 
-import { KeyPair, Key, Nonces, PublicNonces, Signature, NoncePairs } from './types'
+import { Key, Nonces, PublicNonces, Signature, NoncePairs } from './types'
 
 import { _generateL, _aCoefficient, _generatePublicNonces, _multiSigSign, _hashPrivateKey, _sumSigs, _verify, _generatePk, _sign } from './core'
 import { InternalNonces, InternalPublicNonces } from './core/types'
 import { Challenge, FinalPublicNonce, SignatureOutput } from './types/signature'
 
 class Schnorrkel {
-  private nonces: Nonces = {}
+  protected nonces: Nonces = {}
 
   private _setNonce(privateKey: Buffer): string {
     const { publicNonceData, privateNonceData, hash } = _generatePublicNonces(privateKey)
@@ -49,39 +49,6 @@ class Schnorrkel {
     return px
   }
 
-  static fromJson(json: string): Schnorrkel {
-    interface JsonData {
-      nonces: {
-        [hash: string]: {
-          k: string,
-          kTwo: string,
-          kPublic: string,
-          kTwoPublic: string,
-        }
-      }
-    }
-    try {
-      const jsonData = JSON.parse(json) as JsonData
-      const noncesEntries = Object.entries(jsonData.nonces).map(([hash, nonce]) => {
-        return [
-          hash,
-          {
-            k: Key.fromHex(nonce.k),
-            kTwo: Key.fromHex(nonce.kTwo),
-            kPublic: Key.fromHex(nonce.kPublic),
-            kTwoPublic: Key.fromHex(nonce.kTwoPublic),
-          }
-        ]
-      })
-
-      const schnorrkel = new Schnorrkel()
-      schnorrkel.nonces = Object.fromEntries(noncesEntries)
-      return schnorrkel
-    } catch (error) {
-      throw new Error('Invalid JSON')
-    }
-  }
-
   generatePublicNonces(privateKey: Key): PublicNonces {
     const hash = this._setNonce(privateKey.buffer)
     const nonce = this.nonces[hash]
@@ -90,24 +57,6 @@ class Schnorrkel {
       kPublic: nonce.kPublic,
       kTwoPublic: nonce.kTwoPublic,
     }
-  }
-
-  toJson() {
-    const nonces = Object.fromEntries(Object.entries(this.nonces).map(([hash, nonce]) => {
-      return [
-        hash,
-        {
-          k: nonce.k.toHex(),
-          kTwo: nonce.kTwo.toHex(),
-          kPublic: nonce.kPublic.toHex(),
-          kTwoPublic: nonce.kTwoPublic.toHex(),
-        }
-      ]
-    }))
-
-    return JSON.stringify({
-      nonces,
-    })
   }
 
   private clearNonces(privateKey: Key): void {
