+++ src/unsafe-schnorrkel.ts
@@ -1,0 +1,59 @@
+import { Key } from './types'
+
+import { _generateL, _aCoefficient, _generatePublicNonces, _multiSigSign, _hashPrivateKey, _sumSigs, _verify, _generatePk, _sign } from './core'
+import Schnorrkel from './schnorrkel'
+
+class UnsafeSchnorrkel extends Schnorrkel {
+  static fromJson(json: string): UnsafeSchnorrkel {
+    interface JsonData {
+      nonces: {
+        [hash: string]: {
+          k: string,
+          kTwo: string,
+          kPublic: string,
+          kTwoPublic: string,
+        }
+      }
+    }
+    try {
+      const jsonData = JSON.parse(json) as JsonData
+      const noncesEntries = Object.entries(jsonData.nonces).map(([hash, nonce]) => {
+        return [
+          hash,
+          {
+            k: Key.fromHex(nonce.k),
+            kTwo: Key.fromHex(nonce.kTwo),
+            kPublic: Key.fromHex(nonce.kPublic),
+            kTwoPublic: Key.fromHex(nonce.kTwoPublic),
+          }
+        ]
+      })
+
+      const schnorrkel = new UnsafeSchnorrkel()
+      schnorrkel.nonces = Object.fromEntries(noncesEntries)
+      return schnorrkel
+    } catch (error) {
+      throw new Error('Invalid JSON')
+    }
+  }
+
+  toJson() {
+    const nonces = Object.fromEntries(Object.entries(this.nonces).map(([hash, nonce]) => {
+      return [
+        hash,
+        {
+          k: nonce.k.toHex(),
+          kTwo: nonce.kTwo.toHex(),
+          kPublic: nonce.kPublic.toHex(),
+          kTwoPublic: nonce.kTwoPublic.toHex(),
+        }
+      ]
+    }))
+
+    return JSON.stringify({
+      nonces,
+    })
+  }
+}
+
+export default UnsafeSchnorrkel
