+++ tests/core/generateRandomKeys.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import Schnorrkel from '../src/index'
-import { generateRandomKeys } from '../src/core'
+import { generateRandomKeys } from '../../src/core'
 
 describe('testing generateRandomKeys', () => {
   it('should generate key pair', () => {
