+++ tests/schnorrkel/verify.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest'
 
-import Schnorrkel from '../src/index'
-import { _hashPrivateKey, generateRandomKeys } from '../src/core'
+import Schnorrkel from '../../src/index'
+import { _hashPrivateKey, generateRandomKeys } from '../../src/core'
 
 
-describe('testing sumSigs', () => {
+describe('testing verify', () => {
   it('should sum signatures', () => {
     const schnorrkelOne = new Schnorrkel()
     const schnorrkelTwo = new Schnorrkel()
@@ -13,18 +13,20 @@ describe('testing sumSigs', () => {
     const keyPairTwo = generateRandomKeys()
     const publicNoncesOne = schnorrkelOne.generatePublicNonces(keyPairOne.privateKey)
     const publicNoncesTwo = schnorrkelTwo.generatePublicNonces(keyPairTwo.privateKey)
-
+    
     const publicNonces = [publicNoncesOne, publicNoncesTwo]
     const publicKeys = [keyPairOne.publicKey, keyPairTwo.publicKey]
+
+    const combinedPublicKey = Schnorrkel.getCombinedPublicKey(publicKeys)
 
     const msg = 'test message'
     const signatureOne = schnorrkelOne.multiSigSign(keyPairOne.privateKey, msg, publicKeys, publicNonces)
     const signatureTwo = schnorrkelTwo.multiSigSign(keyPairTwo.privateKey, msg, publicKeys, publicNonces)
-
+    
     const signatures = [signatureOne.signature, signatureTwo.signature]
-    const signature = Schnorrkel.sumSigs(signatures)
+    const signaturesSummed = Schnorrkel.sumSigs(signatures)
+    const result = Schnorrkel.verify(signaturesSummed, msg, signatureTwo.finalPublicNonce, combinedPublicKey)
 
-    expect(signature).toBeDefined()
-    expect(signature.buffer).toHaveLength(32)
+    expect(result).toEqual(true)
   })
 })
