+++ tests/unsafe-schnorrkel/generatePublicNonces.test.ts
@@ -1,25 +1,12 @@
 import { describe, expect, it } from 'vitest'
 
-import Schnorrkel from '../src/index'
-import { _hashPrivateKey, generateRandomKeys } from '../src/core'
+import {UnsafeSchnorrkel} from '../../src/index'
+import { _hashPrivateKey, generateRandomKeys } from '../../src/core'
 
 
 describe('testing generatePublicNonces', () => {
-  it('should generate public nonces', () => {
-    const schnorrkel = new Schnorrkel()
-
-    const keyPair = generateRandomKeys()
-    const publicNonces = schnorrkel.generatePublicNonces(keyPair.privateKey)
-
-    expect(publicNonces).toBeDefined()
-    expect(publicNonces.kPublic).toBeDefined()
-    expect(publicNonces.kTwoPublic).toBeDefined()
-    expect(publicNonces.kPublic.buffer).toHaveLength(33)
-    expect(publicNonces.kTwoPublic.buffer).toHaveLength(33)
-  })
-
   it('should overwrite public nonces with same private key', () => {
-    const schnorrkel = new Schnorrkel()
+    const schnorrkel = new UnsafeSchnorrkel()
 
     const keyPair = generateRandomKeys()
     const publicNoncesOne = schnorrkel.generatePublicNonces(keyPair.privateKey)
